iot-weather-solution/modules/weatherModule/requirements.txt
azure-iot-device>=2.0.0
ntplib
requests